make the cache connector a private class inside DictionaryMatcher, add comments
